chore: WIP refactor screenshot-related files
