+++ app/common/renderer/components/SessionInspector/Screenshot/Screenshot.jsx
@@ -1,8 +1,7 @@
 import {Spin} from 'antd';
-import {Fragment, useState} from 'react';
-import {useTranslation} from 'react-i18next';
+import {useState} from 'react';
 
-import {GESTURE_ITEM_STYLES, POINTER_TYPES} from '../../../constants/gestures.js';
+import {POINTER_TYPES} from '../../../constants/gestures.js';
 import {
   DEFAULT_SWIPE,
   DEFAULT_TAP,
@@ -10,14 +9,18 @@ import {
 } from '../../../constants/screenshot.js';
 import {INSPECTOR_TABS} from '../../../constants/session-inspector.js';
 import inspectorStyles from '../SessionInspector.module.css';
+import CoordinatesContainer from './Overlays/CoordinatesContainer.jsx';
 import ElementOverlays from './Overlays/ElementOverlays.jsx';
+import GestureTrail from './Overlays/GestureTrail.jsx';
+import TapSwipeTrail from './Overlays/TapSwipeTrail.jsx';
 import styles from './Screenshot.module.css';
 
 const {POINTER_UP, POINTER_DOWN, PAUSE, POINTER_MOVE} = POINTER_TYPES;
 const {TAP, SELECT, SWIPE, TAP_SWIPE} = SCREENSHOT_INTERACTION_MODE;
 
 /**
- * Shows screenshot of running application and divs that highlight the elements' bounding boxes
+ * Shows the app screenshot along with various overlay elements,
+ * such as divs that highlight the elements' bounding boxes
  */
 const Screenshot = (props) => {
   const {
@@ -31,30 +34,30 @@ const Screenshot = (props) => {
     clearCoordAction,
     scaleRatio,
     selectedTick,
+    tapTickCoordinates,
+    setCoordStart,
+    setCoordEnd,
+    showGesture,
     selectedInspectorTab,
     applyClientMethod,
   } = props;
-  const {t} = useTranslation();
 
   const [x, setX] = useState();
   const [y, setY] = useState();
 
   const handleScreenshotClick = async () => {
-    const {tapTickCoordinates} = props;
     if (selectedTick) {
       await tapTickCoordinates(x, y);
     }
   };
 
   const handleScreenshotDown = async () => {
-    const {setCoordStart} = props;
     if (screenshotInteractionMode === TAP_SWIPE) {
       await setCoordStart(x, y);
     }
   };
 
   const handleScreenshotUp = async () => {
-    const {setCoordEnd} = props;
     if (screenshotInteractionMode === TAP_SWIPE) {
       await setCoordEnd(x, y);
       if (Math.abs(coordStart.x - x) < 5 && Math.abs(coordStart.y - y) < 5) {
@@ -121,45 +124,6 @@ const Screenshot = (props) => {
     await clearCoordAction();
   };
 
-  // retrieve and format gesture for svg drawings
-  const getGestureCoordinates = () => {
-    const {showGesture} = props;
-    const {FILLED, NEW_DASHED, WHOLE, DASHED} = GESTURE_ITEM_STYLES;
-    const defaultTypes = {pointerDown: WHOLE, pointerUp: DASHED};
-
-    if (!showGesture) {
-      return null;
-    }
-    return showGesture.map((pointer) => {
-      // 'type' is used to keep track of the last pointerup/pointerdown move
-      let type = DASHED;
-      const temp = [];
-      for (const tick of pointer.ticks) {
-        if (tick.type === PAUSE) {
-          continue;
-        }
-        const len = temp.length;
-        type = tick.type !== POINTER_MOVE ? defaultTypes[tick.type] : type;
-        if (tick.type === POINTER_MOVE && tick.x !== undefined && tick.y !== undefined) {
-          temp.push({id: tick.id, type, x: tick.x, y: tick.y, color: pointer.color});
-        }
-        if (len === 0) {
-          if (tick.type === POINTER_DOWN) {
-            temp.push({id: tick.id, type: FILLED, x: 0, y: 0, color: pointer.color});
-          }
-        } else {
-          if (tick.type === POINTER_DOWN && temp[len - 1].type === DASHED) {
-            temp[len - 1].type = FILLED;
-          }
-          if (tick.type === POINTER_UP && temp[len - 1].type === WHOLE) {
-            temp[len - 1].type = NEW_DASHED;
-          }
-        }
-      }
-      return temp;
-    });
-  };
-
   // If we're tapping or swiping, show the 'crosshair' cursor style
   const screenshotStyle = {};
   if (screenshotInteractionMode === TAP_SWIPE || selectedTick) {
@@ -169,7 +133,6 @@ const Screenshot = (props) => {
   const screenSrc = isUsingMjpegMode
     ? serverDetails.mjpegScreenshotUrl
     : `data:image/gif;base64,${screenshot}`;
-  const points = getGestureCoordinates();
 
   // Show the screenshot and highlighter rects.
   // Show loading indicator if a method call is in progress, unless using MJPEG mode.
@@ -186,69 +149,20 @@ const Screenshot = (props) => {
           onClick={handleScreenshotClick}
           className={inspectorStyles.screenshotBox}
         >
-          {screenshotInteractionMode !== SELECT && (
-            <div className={styles.coordinatesContainer}>
-              <p>{t('xCoordinate', {x})}</p>
-              <p>{t('yCoordinate', {y})}</p>
-            </div>
-          )}
+          {screenshotInteractionMode !== SELECT && <CoordinatesContainer x={x} y={y} />}
           <img src={screenSrc} id="screenshot" />
           {screenshotInteractionMode === SELECT && <ElementOverlays {...props} />}
           {screenshotInteractionMode === TAP_SWIPE && (
-            <svg className={styles.swipeSvg}>
-              {coordStart && (
-                <circle cx={coordStart.x / scaleRatio} cy={coordStart.y / scaleRatio} r={10} />
-              )}
-              {coordStart && !coordEnd && (
-                <line
-                  x1={coordStart.x / scaleRatio}
-                  y1={coordStart.y / scaleRatio}
-                  x2={x / scaleRatio}
-                  y2={y / scaleRatio}
-                />
-              )}
-              {coordStart && coordEnd && (
-                <line
-                  x1={coordStart.x / scaleRatio}
-                  y1={coordStart.y / scaleRatio}
-                  x2={coordEnd.x / scaleRatio}
-                  y2={coordEnd.y / scaleRatio}
-                />
-              )}
-            </svg>
+            <TapSwipeTrail
+              coordStart={coordStart}
+              coordEnd={coordEnd}
+              x={x}
+              y={y}
+              scaleRatio={scaleRatio}
+            />
           )}
-          {selectedInspectorTab === INSPECTOR_TABS.GESTURES && points && (
-            <svg key="gestureSVG" className={styles.gestureSvg}>
-              {points.map((pointer) =>
-                pointer.map((tick, index) => (
-                  <Fragment key={tick.id}>
-                    {index > 0 && (
-                      <line
-                        className={styles[tick.type]}
-                        key={`${tick.id}.line`}
-                        x1={pointer[index - 1].x / scaleRatio}
-                        y1={pointer[index - 1].y / scaleRatio}
-                        x2={tick.x / scaleRatio}
-                        y2={tick.y / scaleRatio}
-                        style={{stroke: tick.color}}
-                      />
-                    )}
-                    <circle
-                      className={styles[`${tick.type}Circle`]}
-                      key={`${tick.id}.circle`}
-                      cx={tick.x / scaleRatio}
-                      cy={tick.y / scaleRatio}
-                      r={8}
-                      style={
-                        tick.type === GESTURE_ITEM_STYLES.FILLED
-                          ? {fill: tick.color}
-                          : {stroke: tick.color}
-                      }
-                    />
-                  </Fragment>
-                )),
-              )}
-            </svg>
+          {selectedInspectorTab === INSPECTOR_TABS.GESTURES && showGesture && (
+            <GestureTrail gesture={showGesture} scaleRatio={scaleRatio} />
           )}
         </div>
       </div>
